+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - Fix ModMenu icon error.
 - Add French translation.
+- Deprecate the `hoe_required` tag. Use `hoe_never_required` instead. This improves mod compatibility.
 
 [Full Changelog](https://github.com/JamCoreModding/right-click-harvest/compare/3.2.2+1.19.x-1.20.1-fabric...3.2.3+1.19.x-1.20.1-fabric)